style: fix lint errors
