+++ tslint.json
@@ -1,3 +1,6 @@
 {
-  "extends": "tslint-config-standard"
+  "extends": "tslint-config-standard",
+  "linterOptions": {
+    "exclude": "src/**/*.json"
+  }
 }
